+++ src/classes/jsonMetadata.ts
