+++ src/reporter.ts
@@ -1,9 +1,8 @@
 
 const dateformat = require('dateformat')
 // import { getUsers } from "./database";
-import { User } from './user'
-import { Report } from './report';
-
+import { User } from './classes/user'
+import { Report } from './classes/report'
 // const steem = 4.124
 // const spotifyLink = 'https://open.spotify.com/user/1240132288/playlist/4tPNWYmR9liqIS1YgMVClv?si=9KJO_Ap2Th-P4Bivv-37cg'
 // const spotifyImg = 'https://steemitimages.com/DQmVQbXpqHthPR7n2gfcvM4yaUKoez1HRojDNnDA9YqWTpR/image.png'
